Backend/api/requirements.txt
Django==3.1.2
djangorestframework==3.11.0
psycopg2==2.8.6
gunicorn==20.1.0
python-dateutil==2.8.2